Added a test branch for the js.
As to not break anything that currently works, I did as the summary explains.

tl:dr version, nuked some changes to the last millennia and moved them to the newly added test brach. (Yes, the tl:dr version is longer, do you have any problems with that? No? Exactly, didn't think so!)